Refactor for D3 v4
D3 v5 is not yet complete. A number of functions are not working as documented (iterators) and some functions do not yet have any implementations (d3.geoXxx). Version 4 is much more stable.

However, I had used a number of v5 features up until now, so those had to be refactored to work with v4.